task6_1 done.
This script measures and compares the execution time of four different string reversal functions: recursive, iterative, using the built-in 'reversed' method, and slicing. The script defines the functions for string reversal and measures their execution time using the timeit module. It then sorts the functions based on their execution time, identifies the fastest implementation, and calculates the relative slowness of the other functions compared to the fastest one.

Based on the benchmark results, the slicing method (slice_reverse_string) is the fastest and most efficient way to reverse a string among the tested methods. It outperforms the other methods, namely recursive, iterative, and the one using the 'reversed' built-in function. Using the slicing method for string reversal in Python can lead to better code performance.